Isolate generic HDU in a separate object. Keep documenting FitsLib.
